refactor to collect all the js setup stuff back together again
