+++ js-setup/index.js
@@ -1,0 +1,3 @@
+const ComponentInitializer = require('./js/component-initializer');
+
+module.exports = new ComponentInitializer();
