+++ karma.conf.js
@@ -5,7 +5,7 @@ const path = require('path');
 const BowerWebpackPlugin = require('bower-webpack-plugin');
 const componentsToTest = [
 	'test',
-	'layout',
+	'js-setup',
 	'ads',
 	'myft',
 	'tracking',
