--- layout/index.js
@@ -1,3 +1,0 @@
-const JsLoader = require('./js/js-loader');
-
-module.exports = new JsLoader();
