+++ layout/vanilla.html
@@ -14,12 +14,12 @@
 
 		{{#outputBlock 'head'}}{{/outputBlock}}
 		<script>
-			{{>n-ui/layout/partials/ctm}}
+			{{>n-ui/js-loader/templates/ctm}}
 			{{>n-ui/typography/enhance-fonts}}
 		</script>
 	</head>
 	<body class="o-hoverable-on" data-next-is-logged-in="{{@root.anon.userIsLoggedIn}}" data-trackable="page:{{@root.__name}}{{#if @root.trackablePageName}}/{{@root.trackablePageName}}{{/if}}">
 		{{{body}}}
-		{{>n-ui/layout/partials/bootstrapper}}
+		{{>n-ui/js-loader/templates/bootstrapper}}
 	</body>
 </html>
