+++ layout/wrapper.html
@@ -55,7 +55,7 @@
 		{{#outputBlock 'head'}}{{/outputBlock}}
 
 		<script>
-			{{>n-ui/layout/partials/ctm}}
+			{{>n-ui/js-setup/templates/ctm}}
 			{{>n-ui/typography/enhance-fonts}}
 		</script>
 	</head>
@@ -77,7 +77,7 @@
 				{{>n-ui/drawer/template}}
 			</div>
 		</div>
-	{{>n-ui/layout/partials/bootstrapper}}
+	{{>n-ui/js-setup/templates/bootstrapper}}
 	{{>n-ui/tracking/templates/core-analytics}}
 	</body>
 </html>
